+++ modules/data_reading/data_reading.py
@@ -1,10 +1,8 @@
 '''This file contains code reading raw data and do some preprocessing'''
-from multiprocessing import Manager
-import re, json, gc
+import re, json, os
 
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.metrics.pairwise import cosine_similarity
-from datasets import load_dataset
 from bs4 import BeautifulSoup
 import pandas as pd
 import numpy as np
@@ -21,15 +19,11 @@ log.setLevel(logging.ERROR)
 nlp             = spacy.load("en_core_web_sm")
 nlp.add_pipe('sentencizer')
 
+class ContextProcessor():
+    def read_contx(self, id_):
+        with open(f"{PATH['raw_data_dir']}/texts/{id_}.content", 'r', encoding='iso-8859-1') as d_file:
+            return d_file.read()
 
-class DataReading():
-    def __init__(self, split) -> None:
-        self.processed_contx    = {}
-        self.split              = split
-
-    ############################################################
-    # Methods to process context, question and answers
-    ############################################################
     def clean_end(self, text:str):
         text = text.split(' ')
         return ' '.join(text[:-2])
@@ -92,11 +86,20 @@ class DataReading():
 
         return context
 
-    def f_process_contx(self, keys, queue, arg):
-        processed_contx = arg[0]
-        for key in keys:
-            context, kind, start, end = processed_contx[key]
+    def f_process_contx(self, entries, queue, arg):
+        for entry in entries.itertuples():
+            docId   = entry.document_id
 
+            path    = PATH['processed_contx'].replace("[ID]", docId)
+            if check_exist(path):
+                queue.put(1)
+                continue
+
+
+            context = self.read_contx(docId)
+            kind    = entry.kind
+            start   = entry.story_end
+            end     = entry.story_end
 
             ## Extract text from HTML
             context = self.extract_html(context)
@@ -116,17 +119,37 @@ class DataReading():
                 ## Tokenize, remove stopword
                 tokens_ = [tok.text for tok in nlp(sent)
                            if not (tok.is_stop or tok.is_punct)]
-                paras   = np.concatenate((tokens, tokens_))
+                tokens   = np.concatenate((tokens, tokens_))
             len_para = args.seq_len_para - 2
             for i in range(0, len(tokens), len_para):
                 para    = ' '.join(tokens[i: i+len_para])
                 para    = re.sub(r'( |\t){2,}', ' ', para).strip()
 
-                paras   = np.concatenate((paras, para))
+                paras   = np.concatenate((paras, [para]))
 
-            queue.put((key, paras))
+            with open(path, 'w+') as contx_file:
+                json.dump(paras.tolist(), contx_file, indent=2, ensure_ascii=False)
+
+            queue.put(1)
+
+    def trigger_process_contx(self):
+        logging.info(" = Process context.")
+
+        documents   = pd.read_csv(f"{PATH['raw_data_dir']}/documents.csv", header=0, index_col=None)
+
+        for split in ['train', 'test', 'valid']:
+            logging.info(f" = Process context of split: {split}")
+
+            path_dir    = os.path.dirname(PATH['processed_contx'])
+            if not os.path.isdir(path_dir):
+                os.makedirs(path_dir, exist_ok=True)
 
 
+            ParallelHelper(self.f_process_contx, documents[documents["set"] == split],
+                           lambda d, l, h: d.iloc[l:h], args.num_proc).launch()
+
+
+class EntryProcessor():
     def process_ques_ans(self, text):
         """Process question/answers
 
@@ -141,9 +164,6 @@ class DataReading():
 
         return ' '.join(tok)
 
-    ############################################################
-    # Methods to find golden passages given question or answers
-    ############################################################
     def find_golden(self, query, wm: list) -> set:
 
         ## Calculate score of query for each para
@@ -170,6 +190,12 @@ class DataReading():
 
         return goldens
 
+    def read_processed_contx(self, id_):
+        path    = PATH['processed_contx'].replace("[SPLIT]")
+        assert os.path.isfile(path), f"Context with id {id_} not found."
+        with open(path, 'r') as d_file:
+            return json.load(d_file)
+
     def f_process_entry(self, entry):
         """This function is used to run in parallel to read and initially preprocess
         raw documents. Afterward, it puts a dict containing result into queue.
@@ -178,18 +204,15 @@ class DataReading():
             document (dict): a document of dataset
         """
 
-        #########################
-        ## Preprocess context
-        #########################
-        paras   = self.processed_contx[entry['document']['id']]
+        paras   = self.read_processed_contx(entry['document_id'])
 
 
         #########################
         ## Preprocess question and answer
         #########################
-        ques    = self.process_ques_ans(entry['question']['text'])
-        ans1    = self.process_ques_ans(entry['answers'][0]['text'])
-        answ2   = self.process_ques_ans(entry['answers'][1]['text'])
+        ques    = self.process_ques_ans(entry['question'])
+        ans1    = self.process_ques_ans(entry['answer1'])
+        answ2   = self.process_ques_ans(entry['answer2'])
 
 
         #########################
@@ -214,83 +237,48 @@ class DataReading():
         golden_paras_answ   = golden_paras_answ1 | golden_paras_answ2
 
 
-
         return {
-            'doc_id'    : entry['document']['id'],
-            'question'  : ' '.join(entry['question']['tokens']),
-            'answers'   : [' '.join(x['tokens']) for x in entry['answers']],
+            'doc_id'    : entry['document_id'],
+            'question'  : ' '.joind(entry['question_tokenized']),
+            'answers'   : [
+                ' '.join(entry['answer1_tokenized']),
+                ' '.join(entry['answer2_tokenized'])
+            ],
             'En'        : [paras[i] for i in golden_paras_answ],
             'Hn'        : [paras[i] for i in golden_paras_ques]
         }
 
-
-    def trigger_reading_data(self):
-        """ Start reading and processing data
+    def trigger_process_entries(self):
+        """ Start processing pairs of question - context - answer
         """
-        for shard in range(8):
-            #########################
-            # Load shard of dataset
-            #########################
-            ### Need to check whether this shard has already been processed
-            path    = PATH['dataset_para'].replace("[SPLIT]", self.split).replace("[SHARD]", str(shard))
-            if check_exist(path):
-                continue
-            logging.info(f"= Process dataset: {self.split} - shard {shard}")
+        documents   = pd.read_csv(f"{PATH['raw_data_dir']}/qaps.csv", header=0, index_col=None)
 
-            dataset = load_dataset('narrativeqa', split=self.split).shard(8, shard)
+        for split in ['train', 'test', 'valid']:
+            documents   = documents[documents['set'] == split]
+            for shard in range(8):
+                ### Need to check whether this shard has already been processed
+                path    = PATH['dataset'].replace("[SPLIT]", split)\
+                            .replace("[SHARD]", str(shard))
+                if check_exist(path):
+                    continue
 
+                ## Make dir to contain processed files
+                os.makedirs(os.path.dirname(path), exist_ok=True)
 
-
-            #########################
-            # Process contexts first
-            #########################
-            logging.info(f"= Process context in shard {shard}")
-
-            if check_exist(PATH['processed_contx'].replace("[SPLIT]", self.split)):
-                logging.info("=> Backed up processed context file existed. Load it.")
-                with open(PATH['processed_contx'].replace("[SPLIT]", self.split), 'r') as d_file:
-                    self.processed_contx    = json.load(d_file)
+                start_  = len(documents)//8 * shard
+                end_    = start_ + len(documents)//8
 
 
-            # Check whether contexts in this shard have been preprocessed
-            shared_dict = Manager().dict()  # Prepare for multiprocessing
-            for entry in dataset:
-                try:
-                    id_ = entry['document']['id']
-                    _   = self.processed_contx[id_]
-                except KeyError:
-                    shared_dict[id_] = np.array([entry['document']['text'],
-                                                        entry['document']['kind'],
-                                                        entry['document']['start'],
-                                                        entry['document']['end']])
+                list_documents  = [self.f_process_entry(entry)
+                                   for entry in documents.iloc[start_:end_]]
 
-            # Start processing context in parallel
-            list_tuples = ParallelHelper(self.f_process_contx, list(shared_dict.keys()),
-                                        lambda d, l, h: d[l:h],
-                                        args.num_proc, shared_dict).launch()
-
-            # Update self.processed_contx with already processed contexts
-            self.processed_contx.update({it[0]: it[1] for it in list_tuples})
-
-            # Backup processed contexts
-            with open(PATH['processed_contx'].replace("[SPLIT]", self.split), 'w+') as d_file:
-                json.dump(self.processed_contx, d_file, indent=2, ensure_ascii=False)
-                logging.info("=> Backed up processed context.")
-
-            shared_dict = list_tuples = None
-            gc.collect()
-
-
-            #########################
-            # Process each entry of dataset
-            #########################
-            list_documents  = [self.f_process_entry(entry) for entry in dataset]
-
-            save_object(path, pd.DataFrame(list_documents))
+                save_object(path, pd.DataFrame(list_documents))
 
 
 if __name__ == '__main__':
     logging.info("* Reading raw data and decompose into paragraphs")
 
-    for splt in ['test', 'train', 'validation']:
-        DataReading(splt).trigger_reading_data()
+    contx_processor = ContextProcessor()
+    contx_processor.trigger_process_contx()
+
+    entry_processor = EntryProcessor()